Disable mod_deflate
If the Apache server config has mod_deflate output compression enabled, chunked
transfer encoding is used and the Content-Length header is suppressed.

This results in the browser not knowing the download size and not being able to
display download progress.

So here we just disable mod_deflate in case it is active.  If mod_deflate is
not active, setting this environment variable doesn't hurt.

See the comments:
  * http://de.php.net/manual/en/function.flush.php
  * http://php.net/manual/en/function.readfile.php
  * http://php.net/manual/en/function.apache-setenv.php