support non-ascii filenames
Currently, dl-cli bails out on filenames that cannot be encoded by the
ascii codec.

Infer the encoding from the current locale and convert the path and
filename using that.

Note that this assumes that when the filename is passed in as extra parameter
it uses the current encoding, as well. (Currently, no caller does that,
and filename is inferred from the path which comes from the argument.)